+++ include/aws/io/file_utils.h
@@ -47,25 +47,19 @@ int aws_io_translate_and_raise_file_write_error(int error_no);
  * Returns true iff the character is a directory separator on ANY supported platform.
  */
 AWS_IO_API
-bool aws_is_directory_separator(char value);
+bool aws_is_any_directory_separator(char value);
 
 /**
  * Returns the directory separator used by the local platform
  */
 AWS_IO_API
-char aws_get_local_platform_directory_separator(void);
+char aws_get_platform_directory_separator(void);
 
 /**
  * Returns the current user's home directory.
  */
 AWS_IO_API
 struct aws_string *aws_get_home_directory(struct aws_allocator *allocator);
-
-/**
- * Returns the current user's home directory, if it is stored in the HOME environment variable
- */
-AWS_IO_API
-struct aws_string *aws_get_home_directory_environment_value(struct aws_allocator *allocator);
 
 AWS_EXTERN_C_END
