+++ source/posix/file_utils.c
@@ -15,12 +15,22 @@
 
 #include <aws/io/file_utils.h>
 
-char aws_get_local_platform_directory_separator(void) {
+#include <aws/common/environment.h>
+#include <aws/common/string.h>
+
+char aws_get_platform_directory_separator(void) {
     return '/';
 }
+
+AWS_STATIC_STRING_FROM_LITERAL(s_home_env_var, "HOME");
 
 struct aws_string *aws_get_home_directory(struct aws_allocator *allocator) {
 
     /* ToDo: check getpwuid_r if environment check fails */
-    return aws_get_home_directory_environment_value(allocator);
+    struct aws_string *home_env_var_value = NULL;
+    if (aws_get_environment_value(allocator, s_home_env_var, &home_env_var_value) == 0 && home_env_var_value != NULL) {
+        return home_env_var_value;
+    }
+
+    return NULL;
 }
